Изменена структура добавления товаров и категорий

+++ Controllers/ProductController.cs
@@ -15,253 +15,169 @@ namespace Rolled_metal_products.Controllers
     [Authorize(Roles = WC.AdminRole)]
     public class ProductController : Controller
     {
-
-        /*private readonly ApplicationDbContext _context;
-        private readonly IWebHostEnvironment _environment;
-
-        public ProductController(ApplicationDbContext context, IWebHostEnvironment environment) 
-        {
-            _context = context;
-            _environment = environment;
-        }
-        */
         private readonly IProductRepository _prodRepo;
+        private readonly ICategoryRepository _catRepo;
         private readonly IWebHostEnvironment _webHostEnvironment;
-        public ProductController(IProductRepository prodRepo, IWebHostEnvironment webHostEnvironment)
+        public ProductController(IProductRepository prodRepo, ICategoryRepository catRepo, IWebHostEnvironment webHostEnvironment)
         {
             _prodRepo = prodRepo;
+            _catRepo = catRepo;
             _webHostEnvironment = webHostEnvironment;
         }
 
-        /*
-        public IActionResult Index()
-        {
-            IEnumerable<Product> productList = _context.Products.Include(x => x.Category);
-
-            /*foreach (var product in productList) {
-                product.Category = _context.Categories.FirstOrDefault(x => x.Id == product.CategoryId);
-            }*/
-        /*
-            return View(productList);
-        }
-    */
         public IActionResult Index()
         {
             IEnumerable<Product> objList = _prodRepo.GetAll(includeProperties: "Category");
 
-            //foreach(var obj in objList)
-            //{
-            //    obj.Category = _db.Category.FirstOrDefault(u => u.Id == obj.CategoryId);
-            //    obj.ApplicationType = _db.ApplicationType.FirstOrDefault(u => u.Id == obj.ApplicationTypeId);
-            //};
-
             return View(objList);
         }
 
-        //GET - UPSERT
-        public IActionResult Upsert(int? id)
+
+        // GET - CREATE
+        public IActionResult Create(int categoryId)
         {
+            var category = _catRepo.FirstOrDefault(c => c.Id == categoryId, includeProperties: "CategoryParameters");
 
-            //IEnumerable<SelectListItem> CategoryDropDown = _db.Category.Select(i => new SelectListItem
-            //{
-            //    Text = i.Name,
-            //    Value = i.Id.ToString()
-            //});
-
-            ////ViewBag.CategoryDropDown = CategoryDropDown;
-            //ViewData["CategoryDropDown"] = CategoryDropDown;
-
-            //Product product = new Product();
-
-            ProductVM productVM = new ProductVM()
+            if (category == null)
             {
-                Product = new Product(),
-                CategorySelectList = _prodRepo.GetAllDropdownList(WC.CategoryName),
-               // ApplicationTypeSelectList = _prodRepo.GetAllDropdownList(WC.ApplicationTypeName),
+                return NotFound();
+            }
+
+            Product product = new Product();
+            product.CategoryId = categoryId;
+            var model = new CreateProductVM
+            {
+                Product = product,
+                CategoryParameters = category.CategoryParameters.Select(cp => new ProductParameterVM
+                {
+                    CategoryParameterId = cp.Id,
+                    Name = cp.Name
+                }).ToList()
             };
 
-            if (id == null)
-            {
-                //this is for create
-                return View(productVM);
-            }
-            else
-            {
-                productVM.Product = _prodRepo.Find(id.GetValueOrDefault());
-                if (productVM.Product == null)
-                {
-                    return NotFound();
-                }
-                return View(productVM);
-            }
+            return View(model);
         }
-        /*
-        //GET - CREATE
-        
-        public IActionResult Upsert(int? id)
-        {
-            ProductVM productVM = new ProductVM()
-            {
-                Product = new Product(),
-                CategorySelectList = _context.Categories.Select(x => new SelectListItem
-                {
-                    Text = x.Name,
-                    Value = x.Id.ToString()
-                }),
-             
-            };
 
-            if(id == null)
-            {
-                return View(productVM);
-            }
-            {
-                productVM.Product = _context.Products.Find(id); 
-                if(productVM.Product == null)
-                {
-                    return NotFound();
-                }
-            }
-            return View(productVM);
-        }*/
-
-        //POST - UPSERT
+        // POST - CREATE
         [HttpPost]
-        [ValidateAntiForgeryToken]
-        public IActionResult Upsert(ProductVM productVM)
+        public IActionResult Create(CreateProductVM viewModel)
         {
             if (ModelState.IsValid)
             {
                 var files = HttpContext.Request.Form.Files;
                 string webRootPath = _webHostEnvironment.WebRootPath;
 
-                if (productVM.Product.Id == 0)
+                string upload = webRootPath + WC.ImagePath;
+                string fileName = Guid.NewGuid().ToString();
+                string extension = Path.GetExtension(files[0].FileName);
+
+                using (var fileStream = new FileStream(Path.Combine(upload, fileName + extension), FileMode.Create))
                 {
-                    //Creating
-                    string upload = webRootPath + WC.ImagePath;
-                    string fileName = Guid.NewGuid().ToString();
-                    string extension = Path.GetExtension(files[0].FileName);
-
-                    using (var fileStream = new FileStream(Path.Combine(upload, fileName + extension), FileMode.Create))
-                    {
-                        files[0].CopyTo(fileStream);
-                    }
-
-                    productVM.Product.ImageName = fileName + extension;
-
-                    _prodRepo.Add(productVM.Product);
+                    files[0].CopyTo(fileStream);
                 }
-                else
+
+                viewModel.Product.ImageName = fileName + extension;
+
+                var product = viewModel.Product;
+
+                product.ProductParameters = viewModel.CategoryParameters.Select(pp => new ProductParameter
                 {
-                    //updating
-                    var objFromDb = _prodRepo.FirstOrDefault(u => u.Id == productVM.Product.Id, isTracking: false);
+                    ProductId = viewModel.Product.Id,
+                    Name = pp.Name,
+                    Value = pp.Value
+                }).ToList();
 
-                    if (files.Count > 0)
-                    {
-                        string upload = webRootPath + WC.ImagePath;
-                        string fileName = Guid.NewGuid().ToString();
-                        string extension = Path.GetExtension(files[0].FileName);
-
-                        var oldFile = Path.Combine(upload, objFromDb.ImageName);
-
-                        if (System.IO.File.Exists(oldFile))
-                        {
-                            System.IO.File.Delete(oldFile);
-                        }
-
-                        using (var fileStream = new FileStream(Path.Combine(upload, fileName + extension), FileMode.Create))
-                        {
-                            files[0].CopyTo(fileStream);
-                        }
-
-                        productVM.Product.ImageName = fileName + extension;
-                    }
-                    else
-                    {
-                        productVM.Product.ImageName = objFromDb.ImageName;
-                    }
-                    _prodRepo.Update(productVM.Product);
-                }
-                TempData[WC.Success] = "Action completed successfully";
-
+                _prodRepo.Add(product);
                 _prodRepo.Save();
-                return RedirectToAction("Index");
+
+                return RedirectToAction("Details", "Category", new { id = product.CategoryId });
+                //return RedirectToAction("","Index");
             }
-            productVM.CategorySelectList = _prodRepo.GetAllDropdownList(WC.CategoryName);
-           // productVM.ApplicationTypeSelectList = _prodRepo.GetAllDropdownList(WC.ApplicationTypeName);
 
-            return View(productVM);
-
+            return View(viewModel);
         }
-        /*
-        //POST - CREATE
+
+        // GET - EDIT 
+        [HttpGet]
+        public IActionResult Edit(int id)
+        {
+            var product = _prodRepo.FirstOrDefault(p => p.Id == id, includeProperties: "ProductParameters");
+
+            if (product == null)
+            {
+                return NotFound();
+            }
+
+            var categoryParameters = product.ProductParameters
+                .Select(pp => new ProductParameterVM
+                {
+                    Name = pp.Name,
+                    Value = pp.Value
+                }).ToList();
+
+            var viewModel = new CreateProductVM
+            {
+                Product = product,
+                CategoryParameters = categoryParameters
+            };
+
+            return View(viewModel);
+        }
+
+        // POST - EDIT
         [HttpPost]
-        [ValidateAntiForgeryToken]
-        public IActionResult Upsert(ProductVM productVM)
+        public IActionResult Edit(CreateProductVM viewModel)
         {
             if (ModelState.IsValid)
             {
-                var files = HttpContext.Request.Form.Files;
-                string webRootPath = _environment.WebRootPath;
+                var productFromDb = _prodRepo.FirstOrDefault(u => u.Id == viewModel.Product.Id, isTracking: false);
 
-                if (productVM.Product.Id == 0)
+                var files = HttpContext.Request.Form.Files;
+                string webRootPath = _webHostEnvironment.WebRootPath;
+
+                if (files.Count > 0)
                 {
                     string upload = webRootPath + WC.ImagePath;
                     string fileName = Guid.NewGuid().ToString();
                     string extension = Path.GetExtension(files[0].FileName);
+
+                    var oldFile = Path.Combine(upload, productFromDb.ImageName);
+
+                    if (System.IO.File.Exists(oldFile))
+                    {
+                        System.IO.File.Delete(oldFile);
+                    }
 
                     using (var fileStream = new FileStream(Path.Combine(upload, fileName + extension), FileMode.Create))
                     {
                         files[0].CopyTo(fileStream);
                     }
 
-                    productVM.Product.ImageName = fileName + extension;
-                    _context.Products.Add(productVM.Product);
+                    viewModel.Product.ImageName = fileName + extension;
                 }
-                else 
+                else
                 {
-                    var productFromDb = _context.Products.AsNoTracking().FirstOrDefault(x => x.Id == productVM.Product.Id);
-
-
-                    if (files.Count > 0)
-                    {
-                        string upload = webRootPath + WC.ImagePath;
-                        string fileName = Guid.NewGuid().ToString();
-                        string extension = Path.GetExtension(files[0].FileName);
-
-                        var oldFile = Path.Combine(upload, productFromDb.ImageName);
-
-                        if (System.IO.File.Exists(oldFile))
-                        {
-                            System.IO.File.Delete(oldFile);
-                        }
-
-                        using (var fileStream = new FileStream(Path.Combine(upload, fileName + extension), FileMode.Create))
-                        {
-                            files[0].CopyTo(fileStream);
-                        }
-
-                        productVM.Product.ImageName = fileName + extension;
-                    }
-                    else 
-                    {
-                        productVM.Product.ImageName = productFromDb.ImageName;
-                    }
-                    _context.Products.Update(productVM.Product);
+                    viewModel.Product.ImageName = productFromDb.ImageName;
                 }
-                _context.SaveChanges();
-                return RedirectToAction("Index");
+
+                _prodRepo.DeleteExistingParameters(viewModel.Product.Id);
+
+                var product = viewModel.Product;
+
+                product.ProductParameters = viewModel.CategoryParameters.Select(pp => new ProductParameter
+                {
+                    ProductId = viewModel.Product.Id,
+                    Name = pp.Name,
+                    Value = pp.Value
+                }).ToList();
+
+                _prodRepo.Update(product);
+                _prodRepo.Save();
+
+                return RedirectToAction("Details", "Category", new { id = product.CategoryId });
             }
-
-            productVM.CategorySelectList = _context.Categories.Select(x => new SelectListItem
-            {
-                Text = x.Name,
-                Value = x.Id.ToString()
-            });
-
-            return View(productVM);
+            return View(viewModel);
         }
-        */
 
         //GET - DELETE
         public IActionResult Delete(int? id)
@@ -279,62 +195,13 @@ namespace Rolled_metal_products.Controllers
 
             return View(product);
         }
-        /*
-        //GET - DELETE
-        public IActionResult Delete(int? id)
-        {
-            if (id == null || id == 0)
-            {
-                return NotFound();
-            }
-            Product product = _context.Products.Include(u => u.Category).FirstOrDefault(u => u.Id == id);
-            if (product == null)
-            {
-                return NotFound();
-            }
-            return View(product);
-        }*/
-        /*
+    
         //POST - DELETE
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public IActionResult DeletePost(int? id)
         {
-            if (id == null || id == 0)
-            {
-                return NotFound();
-            }
-
-            var product = _context.Products.Find(id);
-
-            if (product == null)
-            {
-                return NotFound();
-            }
-
-            string webRootPath = _environment.WebRootPath;
-            string upload = webRootPath + WC.ImagePath;
-            var oldFile = Path.Combine(upload, product.ImageName);
-
-            if (System.IO.File.Exists(oldFile))
-            {
-                System.IO.File.Delete(oldFile);
-            }
-
-            if (ModelState.IsValid)
-            {
-                _context.Remove(product);
-                _context.SaveChanges();
-                return RedirectToAction("Index");
-            }
-            return View(product);
-        }*/
-        //POST - DELETE
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public IActionResult DeletePost(int? id)
-        {
-            var obj = _prodRepo.Find(id.GetValueOrDefault());
+            var obj = _prodRepo.FirstOrDefault(x => x.Id == id, includeProperties: "ProductParameters");
             if (obj == null)
             {
                 return NotFound();
@@ -348,13 +215,14 @@ namespace Rolled_metal_products.Controllers
                 System.IO.File.Delete(oldFile);
             }
 
-
-            _prodRepo.Remove(obj);
-            _prodRepo.Save();
-            TempData[WC.Success] = "Action completed successfully";
-            return RedirectToAction("Index");
-
-
+            if (ModelState.IsValid)
+            {
+                _prodRepo.Remove(obj);
+                _prodRepo.Save();
+                TempData[WC.Success] = "Action completed successfully";
+                return RedirectToAction("Details", "Category", new { id = obj.CategoryId });
+            }
+            return View(obj);
         }
     }
 }
